requirements.txt
# Core Django and database
Django==4.2.16
psycopg2-binary==2.9.10

# Authentication and security
PyJWT==2.8.0
bcrypt==4.1.3
passlib[bcrypt]==1.7.4
cryptography==46.0.2

# Development and testing
pytest==8.4.2
pytest-django==4.9.0
mypy==1.18.2
django-stubs==5.1.2

# Code formatting and linting
black==25.9.0
pylint==3.3.8
isort==6.1.0

# API and HTTP
djangorestframework==3.15.2
requests==2.32.3

# Environment management
python-dotenv==1.0.1

# Utilities
python-dateutil==2.9.0